# Add specific BuildScoreProvider for diversity to avoid extra encoding…