… and decoding of nodes

For a JMH just benchmarking the diversity calculation this is a huge win

Before:
```
PQDistanceCalculationBenchmark.distanceCalculation          0         1536           100          10000  avgt    5   418.095 ±  4.628  ms/op
PQDistanceCalculationBenchmark.distanceCalculation         16         1536           100          10000  avgt    5   940.306 ±  2.556  ms/op
PQDistanceCalculationBenchmark.distanceCalculation         64         1536           100          10000  avgt    5  1214.263 ± 70.999  ms/op
PQDistanceCalculationBenchmark.distanceCalculation        192         1536           100          10000  avgt    5  2019.785 ± 67.312  ms/op
```

```
Benchmark                                           (M)  (dimension)  (queryCount)  (vectorCount)  Mode  Cnt    Score   Error  Units
PQDistanceCalculationBenchmark.distanceCalculation    0         1536           100          10000  avgt    5  417.770 ± 3.297  ms/op
PQDistanceCalculationBenchmark.distanceCalculation   16         1536           100          10000  avgt    5  261.959 ± 3.048  ms/op
PQDistanceCalculationBenchmark.distanceCalculation   64         1536           100          10000  avgt    5  376.058 ± 3.726  ms/op
PQDistanceCalculationBenchmark.distanceCalculation  192         1536           100          10000  avgt    5  666.985 ± 8.505  ms/op

```


For actual Graph Build using PQ diversity this is more like 25% boost

Before
```
Benchmark                                                    (numBaseVectors)  (numberOfPQSubspaces)  (originalDimension)  Mode  Cnt      Score      Error  Units
IndexConstructionWithRandomSetBenchmark.buildIndexBenchmark             10000                     48                  384  avgt    3   2998.352 ±  292.077  ms/op
IndexConstructionWithRandomSetBenchmark.buildIndexBenchmark            100000                     48                  384  avgt    3  30923.062 ± 1689.046  ms/op
```

After
```
Benchmark                                                    (numBaseVectors)  (numberOfPQSubspaces)  (originalDimension)  Mode  Cnt      Score      Error  Units
IndexConstructionWithRandomSetBenchmark.buildIndexBenchmark             10000                     48                  384  avgt    3   2370.760 ±  230.455  ms/op
IndexConstructionWithRandomSetBenchmark.buildIndexBenchmark            100000                     48                  384  avgt    3  25302.423 ± 2256.221  ms/op

```